-added impression model pred in CJCIE -fixed some bugs in search in CJCIE -added "better" imp predictions in IETest (will give misleading results=])
git-svn-id: svn://cs2955/var/svn/client/trunk@2163 729a0f52-b9d0-41be-9dfd-8227063685e9

+++ src/models/queryanalyzer/iep/IEResult.java
@@ -3,28 +3,39 @@ package models.queryanalyzer.iep;
 import java.util.Arrays;
 
 public class IEResult {
-	private int[] _order;
-	private int[] _sol; //numImps each agent saw
-	private int[] _slotImpr; //numImps in each slot
-	private int _obj;
-	
-	public IEResult(int obj, int[] sol, int[] order, int[] slotImpr){
-		_obj = obj;
-		_sol = sol;
-		_order = order;
-		_slotImpr = slotImpr;
-	}
-	
-	public int getObj() {return _obj;}
-	public int[] getSol() {return _sol;}
-	public int[] getOrder() {return _order;}
-	public int[] getSlotImpressions() {return _slotImpr;}
-	
-	public String toString() {
-		StringBuffer sb = new StringBuffer();
-		sb.append("Result:\n\tobj=" + _obj + "\n\tsol=" + Arrays.toString(_sol) + "\n\torder=" + Arrays.toString(_order) + "\n\tslotImpr=" + Arrays.toString(_slotImpr));
-		//sb.append("Result:\tobj=" + _obj + "\tsol=" + Arrays.toString(_sol));
-		return sb.toString();
-		
-	}
+   private int[] _order;
+   private int[] _sol; //numImps each agent saw
+   private int[] _slotImpr; //numImps in each slot
+   private double _obj;
+
+   public IEResult(double obj, int[] sol, int[] order, int[] slotImpr) {
+      _obj = obj;
+      _sol = sol;
+      _order = order;
+      _slotImpr = slotImpr;
+   }
+
+   public double getObj() {
+      return _obj;
+   }
+
+   public int[] getSol() {
+      return _sol;
+   }
+
+   public int[] getOrder() {
+      return _order;
+   }
+
+   public int[] getSlotImpressions() {
+      return _slotImpr;
+   }
+
+   public String toString() {
+      StringBuffer sb = new StringBuffer();
+      sb.append("Result:\n\tobj=" + _obj + "\n\tsol=" + Arrays.toString(_sol) + "\n\torder=" + Arrays.toString(_order) + "\n\tslotImpr=" + Arrays.toString(_slotImpr));
+      //sb.append("Result:\tobj=" + _obj + "\tsol=" + Arrays.toString(_sol));
+      return sb.toString();
+
+   }
 }
